debugging server-client

+++ SOURCE/client.cpp
@@ -89,12 +89,12 @@ void receiveMessage_C(int socketFD, char* buffer)
 	// 	exit(0);
 	// }
 
-		size_t len = sizeof(buffer);
+		int len = sizeof(buffer);
 		char *p = buffer;
-		ssize_t n;
+		int n;
 		while ( len > 0 && (n=recv(socketFD,p,len,0)) > 0 ) {
 		  p += n;
-		  len =- (size_t)n;
+		  len =- n;
 		}
 		if ( len > 0 || n < 0 ) {
 		  // oops, something went wrong
